requirements.txt
requests==2.23.0
beautifulsoup4==4.9.0
xlsxwriter==1.2.3